requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10